resourceloader: Give SkinModule 'features' option an extensible default
The previous method of using an array was restrictive in that core
could not add new features automatically to skins.

This change allows an opt-out policy as well as the existing
opt-in policy. Wikimedia deployed skins will soon be adapted to take
the opt-in policy.

Bug: T252774
Bug: T271441
Change-Id: I517f9f80760348da589852cda53cbd0c14231f86
(cherry picked from commit 6ab14857b972eea9a0fb94f8d50e5d098b00be51)