Add a link from a deleted page to that page's logs
For pages that exist, there is a link to view that page's logs at
the top of the "history" screen. However, there is no easy way to
get from a deleted page to that page's logs. This patch adds such a
link, which will appear after the "View or restore N deleted edits?"
link for any user with the `deletedhistory` right.

Bug: T261087
Change-Id: I4f11a400360bab3045ab5de86d1270e5c2c6b10a