Fix base path handling for MessagePosterModule registration
It was done in a funny way to allow extensions to provide paths that
are relative to a different base path than the core ones. This way
results in wrong paths on Windows. Use the ResourceLoaderFilePath
class instead, which I wrote for this exact purpose :D

With this approach, it should also be possible to handle remote paths
correctly, and remove the `'debugRaw' => false` hack in ServiceWiring,
but I don't entirely understand how that is handled in the extension
registration code.

Bug: T264965
Change-Id: Id501a10d87a60f8c71df49820e1fb866adc12411