Don't assume that warnings array will include 'code' key
This code is explicitly in place to allow continuing an upload
in the case of (recoverable) warning. The `code` looks to only
be used as part of the failure handling, and the lack of that
failure-related key should not cause the upload process to halt
if not all chunks have been uploaded yet.

Change-Id: Ib9fad816c54662aa2642821195756e0f2828919b