Fix English/*nix specific error messages in FSFileBackend
FileBackend::getFileList() is supposed to return an empty array when the
specified path does not exist, however this was broken on Windows due to
a difference in the exception message raised by DirectoryIterator.

Unit tests and code inspection showed that there were several other
instances of unmatched Windows error messages, which I also fixed. Now
FileBackendIntegrationTest passes on Windows.

It was also broken when $wgShellLocale is set to a non-English locale.
Recommend against this in the default LocalSettings.php. Work around
the issue using socket_strerror() if available. I was able to confirm
the bug and the fix using the existing tests.

Bug: T265778
Change-Id: I2b4cfbb1554d9e05cce53513f4b474064fb57de5
(cherry picked from commit 5b1c7fc157da6539cbdd0b69b177fc1e484365b3)