wfAssembleUrl: Handle empty query field in URL bits
parse_url() on PHP >= 8 will include an empty 'query' bit for input URLs that
end in '?' and have no query string, such as 'http://www.example.com/foo?'.
This causes the parser test for T4372 to break. Accordingly, fix wfAssembleUrl()
to not generate a query string if the input query field is empty.

Bug: T268852
Change-Id: Ief9c15cc8fe0fcae07d2cc3120db9ca95573324c