Don't rely on implicit string->int cast in comparison
PHP8 came with saner string to number comparisons.

In PHP8, the empty string "" will not be cast to integer before
the comparison. And so literally 0 == "" is false. Previously it
will be cast, so the eventual comparison was 0 == 0, which is true.

Bug: T272250
Change-Id: Ic280467e520840d45cbd172476252c6de5164100
(cherry picked from commit 995e2a0eed31687e1af34929ee8cdabb854239cf)