Pass linkable user name in Special:Preferences "Account data" help text
For the optional message in the "Account data" section, introduced in
d4357d621fa6fa09bb6fb09c57ee21331f92f260.

Passes the user name escaped with underscores, so that links can be
built correctly in the message translation.

This means we have to expand yet further the list of User methods called
in ApiOptionsTest, but so be it.

Bug: T272412
Change-Id: Ia1c135144763db9e92f1b555f009229bd1a9d3c8