Add Title::getId() as an alias for ::getArticleId()
Title::getId() was introduced in Iaed4871e0d32c6. It was
documented to be @since 1.35, which actually isn't the case.
There are now extensions who are documented to support 1.35,
but they actually aren't.

Adding Title::getId() as an alias for ::getArticleId() in 1.35
seems to be the simplest solution to this.

It gets an optional $wikiId parameter in order to match the
structure of the method in the new releases.

Bug: T303871
Change-Id: I6f249df8171e670e19812aa2adf517883c84c0f6